police movements improved.
addition to dijkstra: you can choose if you want to get the sortest path
or the cheapest path.

now when a police wants to reach a node, it calculates the turns it will
take to get there, using "shortestPath" and the "cheapestPath". the
police will choose the path that takes less  turns to get there.