thieves movement improved.
there has been some changes, which splits up thieves if they stick
together.
and if all neighbors of a thief have police in them, or have a police in
neighboring of them, the thief will move to one one the neighbors that
has a police in neighboring of it, the thief might get caught or may
not, but if it does not move, it will get caught.